backend/requirements.txt
# --- Core API ---
fastapi>=0.118.0
uvicorn[standard]>=0.37.0
python-multipart>=0.0.20
python-dotenv>=1.1.1    # Useful for local .env loading

# --- AI & Image Processing ---
torch                   # The Deep Learning Model
torchvision             # Transforms (ResNet)
numpy>=2.0.0            # Math for scores
pillow>=11.3.0          # Image loading & resizing

# --- Database ---
sqlalchemy>=2.0.43      # ORM
psycopg2-binary>=2.9.9  # Postgres Adapter
pydantic>=2.11.10       # Data Validation

# --- Testing
pytest 
pytest-cov 
pytest-asyncio 
httpx